fastapi_template/config.py
import os

from dotenv import load_dotenv
from pydantic import BaseSettings

BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
load_dotenv(os.path.join(BASE_DIR, ".env"))


class Settings(BaseSettings):
    """Global config."""

    # RUN
    HOST = os.getenv("HOST", "127.0.0.1")
    PORT = int(os.getenv("PORT", 8000))
    LOG_LEVEL = os.getenv("LOG_LEVEL", "debug")
    RELOAD = bool(os.getenv("RELOAD", True))
    DEBUG = bool(os.getenv("DEBUG", True))


settings = Settings()
