pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"
packages = [{include = "fastapi_template"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.22.0"
fastapi = "^0.97.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
